improvements on run performed organization
