+++ python_search/entry_runner.py
@@ -8,7 +8,7 @@ from python_search.apps.notification_ui import send_notification
 from python_search.config import PythonSearchConfiguration
 from python_search.context import Context
 from python_search.events.run_performed import RunPerformed
-from python_search.events.run_performed.client import LogRunPerformedClient
+from python_search.events.run_performed.writer import LogRunPerformedClient
 from python_search.interpreter.cmd import CmdInterpreter
 from python_search.interpreter.interpreter_matcher import InterpreterMatcher
 from python_search.logger import setup_run_key_logger
