+++ python_search/events/run_performed/__init__.py
@@ -1,39 +1,3 @@
 from __future__ import annotations
 
-from typing import Optional
-
-from pydantic import BaseModel
-
-from python_search.data_collector import GenericDataCollector
-
-
-class RunPerformedWriter:
-    def write(self, event: RunPerformed):
-        import datetime
-
-        event.timestamp = str(datetime.datetime.now(datetime.timezone.utc).timestamp())
-
-        return GenericDataCollector().write(
-            data=event.__dict__, table_name="searches_performed"
-        )
-
-
-class RunPerformed(BaseModel):
-    """
-    Main event of the application.
-    Identifies a search being executed
-    """
-
-    # name of the entry matched
-    key: str
-    # for when a query was typed by the user
-    # @todo rename to something more meaningful
-    query_input: str
-    # for when it is started from a shortcut
-    shortcut: str
-    # unix timestamp
-    timestamp: Optional[str] = None
-
-    @staticmethod
-    def get_schema():
-        return "key string, query_input string, shortcut string"
+from python_search.events.run_performed.entity import RunPerformed
