+++ python_search/events/run_performed/writer.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from python_search.data_collector import GenericDataCollector
 from python_search.events.run_performed import RunPerformed
 
 
@@ -16,3 +17,14 @@ class LogRunPerformedClient:
             return result
         except BaseException as e:
             print(f"Logging results failed, reason: {e}")
+
+
+class RunPerformedWriter:
+    def write(self, event: RunPerformed):
+        import datetime
+
+        event.timestamp = str(datetime.datetime.now(datetime.timezone.utc).timestamp())
+
+        return GenericDataCollector().write(
+            data=event.__dict__, table_name="searches_performed"
+        )
